CHG: Improve Dask communication
