+++ syncopy/shared/kwarg_decorators.py
@@ -16,7 +16,7 @@ from syncopy.shared.errors import (SPYTypeError, SPYValueError,
                                    SPYError, SPYWarning, SPYInfo)
 from syncopy.shared.tools import StructDict
 from syncopy.shared.metadata import h5_add_metadata, parse_cF_returns
-from .tools import check_slurm_available
+from .dask_helpers import check_slurm_available, check_workers_available
 import syncopy as spy
 
 __all__ = []
@@ -474,7 +474,10 @@ def detect_parallel_client(func):
         # this needs explicit `parallel=True`.
         if parallel is None:
             try:
-                dd.get_client()
+                client = dd.get_client()
+                check_workers_available(client.cluster)
+                msg = f"..attaching to running Dask client:\n{client}"
+                SPYInfo(msg)
                 parallel = True
             except ValueError:
                 parallel = False
@@ -486,22 +489,26 @@ def detect_parallel_client(func):
             # if already one cluster is reachable do nothing
             try:
                 client = dd.get_client()
+                check_workers_available(client.cluster)
                 msg = f"..attaching to running Dask client:\n{client}"
                 SPYInfo(msg)
             except ValueError:
-                # we are on a HPC but ACME is missing, LocalCluster still got created
-                # but a warning is issued
+                # we are on a HPC but ACME and Dask client are missing,
+                # LocalCluster still gets created
+                slurm_msg = ""
                 if has_slurm and not spy.__acme__:
-                    msg = ("We are apparently on a slurm cluster but could not find a Dask client.\n"
-                           "Syncopy does not provide an "
-                           "automatic Dask SLURMCluster on its own!"
-                           "\nPlease consider using ACME (https://github.com/esi-neuroscience/acme)"
-                           "\nor configure your own cluster via `dask_jobqueue.SLURMCluster()`"
-                           "\n\nCreating a LocalCluster as fallback.."
+                    slurm_msg = ("We are apparently on a slurm cluster but\n"
+                                 "Syncopy could not find a Dask client.\n"
+                                 "Syncopy does not provide an "
+                                 "automatic Dask SLURMCluster on its own!"
+                                 "\nPlease consider using ACME (https://github.com/esi-neuroscience/acme)"
+                                 "\nor configure your own cluster via `dask_jobqueue.SLURMCluster()`"
+                                 "\n\nCreating a LocalCluster as fallback.."
                            )
-                    SPYWarning(msg)
+                    SPYWarning(slurm_msg)
 
-                # spawn fallback local cluster
+                # -- spawn fallback local cluster --
+
                 cluster = dd.LocalCluster()
                 # attaches to local cluster residing in global namespace
                 dd.Client(cluster)
@@ -521,6 +528,9 @@ def detect_parallel_client(func):
             dd.get_client().close()
             # and kill
             cluster.close()
+        # print again in case it got drowned
+        if slurm_msg:
+            SPYWarning(slurm_msg)
 
         return results
